fix: CI빌드 에러
