+++ src/app/api/users/index.ts
@@ -130,3 +130,15 @@ export const getExpiredMySolutions = async ({
 
   return response;
 };
+
+export const deleteSignOut = async (accessToken: string) => {
+  const response = await kyInstance
+    .delete("api/user/logout", {
+      headers: {
+        Authorization: `Bearer ${accessToken}`,
+      },
+    })
+    .json();
+
+  return response;
+};
