--- src/app/api/users/signOut.ts
@@ -1,13 +1,0 @@
-import { kyInstance } from "@/api";
-
-export const deleteSignOut = async (accessToken: string) => {
-  const response = await kyInstance
-    .delete("api/user/logout", {
-      headers: {
-        Authorization: `Bearer ${accessToken}`,
-      },
-    })
-    .json();
-
-  return response;
-};
